Refactored Code from uart.c to new file sfu_cmds.c. Moved code that takes commands from the terminal and returns information such as 'get heap' , 'help' etc
